requirements.txt
opencv-python-headless
beautifulsoup4
huggingface-hub==0.23.4
requests
langchain
mysql-connector-python
openai
langchain-openai
langchain-community
chromadb
python-dotenv
torch>=1.7
torchvision>=0.8
pandas
flask
line-bot-sdk
tf-keras
transformers==4.41.2
sentence_transformers==2.2.2